Reproducibility/RDM changes
Added an optional section about definition of reproducibility. 
Added a description of what we are currently providing with workshops/programs. 
Added comments about promotion and advertising
Added paragraph related to RDM partnerships. I haven't added anything about partnering with UITS, mainly through RCGC and the Research Computing unit. Not as close a relationship as with RDI. 
Also still need to add a section about the future and offering Drop-in hours/short presentations coming this fall.